refactor: restructure codebase by moving context package
